feat: refactor file structure, change imports
