+++ components/hooks/TransitionDemo.tsx
@@ -1,12 +1,12 @@
-"use client";
+'use client';
 
-import React from "react";
+import React from 'react';
 
-import { hookData } from "@/data/hooks";
-import { ExampleBlock } from "../shared";
+import { hookData } from '@/data/react/hooks';
+import { ExampleBlock } from '../shared';
 
 export function TransitionDemo() {
-  const [input, setInput] = React.useState("");
+  const [input, setInput] = React.useState('');
   const [list, setList] = React.useState<string[]>([]);
   const [isPending, startTransition] = React.useTransition();
 
@@ -44,11 +44,7 @@ export function TransitionDemo() {
           placeholder="Start typing..."
           className="px-3 py-1 rounded bg-gray-900 border border-gray-600 text-white w-full mb-2"
         />
-        {isPending && (
-          <p className="text-yellow-400 text-sm mb-2">
-            Rendering large list...
-          </p>
-        )}
+        {isPending && <p className="text-yellow-400 text-sm mb-2">Rendering large list...</p>}
         <div className="h-32 overflow-auto bg-gray-950 text-xs p-2 rounded border border-gray-700">
           {list.slice(0, 20).map((item, i) => (
             <div key={i}>{item}</div>
